requirements.txt
alembic==1.15.1
APScheduler==3.11.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-JWT-Extended==4.7.1
Flask-Logging==0.1.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
